Codigo del formulario prestamos actualizado

+++ Entidades/Préstamos.cs
@@ -11,10 +11,10 @@ namespace Entidades
         public int IdPrestamo { get; set; }
         public int IdCliente { get; set; }
         public int IdUsuario { get; set; }
-        public int MontoCapital { get; set; }
+        public decimal MontoCapital { get; set; } // cambio a decimal
         public int PlazoMeses { get; set; }
         public decimal TasaInteresAplicada { get; set; }
-        public int MontoTotal { get; set; }
+        public decimal MontoTotal { get; set; } // cambio a decimal
         public string Garantia { get; set; }
         public DateTime Fecha { get; set; }
     }
